ENGOPS-1813-bitbucket-7-compatibility implement parallel dryRun merges with post sorting.  Refactor dryRunMerge method to not mutate mcd object.  Clean up exception handling.
